+++ FirebasePerformance/Sources/Configurations/FPRRemoteConfigFlags+Private.h
@@ -24,7 +24,7 @@ static NSString *const kFPRConfigPrefix = @"com.fireperf";
 static NSInteger const kFPRConfigFetchIntervalInSeconds = 12 * 60 * 60;
 
 /** Interval after which the configurations can be fetched. Specified in seconds. */
-static NSInteger const kFPRConfigAppStartDelayInSeconds = 1 * 30;
+static NSInteger const kFPRMinAppStartConfigFetchDelayInSeconds = 5;
 
 /** This extension should only be used for testing. */
 @interface FPRRemoteConfigFlags ()
@@ -33,10 +33,19 @@ static NSInteger const kFPRConfigAppStartDelayInSeconds = 1 * 30;
 @property(nonatomic) FIRRemoteConfig *fprRemoteConfig;
 
 /** @brief Last activated time of the configurations. */
-@property(atomic) NSDate *lastFetchedTime;
+@property(atomic, nullable) NSDate *lastFetchedTime;
 
 /** @brief User defaults used for caching. */
 @property(nonatomic) NSUserDefaults *userDefaults;
+
+/** @brief Last activated time of the configurations. */
+@property(nonatomic) NSDate *applicationStartTime;
+
+/** @brief Number of seconds delayed until the first config is made during app start. */
+@property(nonatomic) NSTimeInterval appStartConfigFetchDelayInSeconds;
+
+/** @brief Status of the last remote config fetch. */
+@property(nonatomic) FIRRemoteConfigFetchStatus lastFetchStatus;
 
 /**
  * Creates an instance of FPRRemoteConfigFlags.