+++ FirebasePerformance/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Pending
+* Create a random number of delay for remote config fetch during app starts.
+
 # Version 8.6.1
 * Fix the case where the event were dropped for missing a critical field in the event.